mcp_server_plugin/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
aiomysql==0.2.0
sqlalchemy==2.0.25
google-generativeai==0.3.2
redis==5.0.1
httpx==0.26.0
python-multipart==0.0.6
cryptography==41.0.7
